pyproject.toml
[project]
name = "bdproject"
version = "0.1.0"
requires-python = "==3.13.*"
dependencies = [
    "confluent-kafka>=2.11.1",
    "contextily>=1.6.2",
    "crick>=0.0.8",
    "cryptography>=45.0.3",
    "dask-geopandas>=0.5.0",
    "dask-jobqueue>=0.9.0",
    "dask[complete]>=2025.5.1",
    "distributed>=2025.5.1",
    "duckdb>=1.3.0",
    "geopandas>=1.1.0",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.7",
    "itables>=2.4.0",
    "jupysql>=0.11.1",
    "jupyter-server>=2.16.0",
    "matplotlib>=3.10.3",
    "memray>=1.17.2",
    "polars>=1.30.0",
    "pyarrow>=20.0.0",
    "requests>=2.32.3",
    "seaborn>=0.13.2",
    "tables>=3.10.2",
    "tqdm>=4.67.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pyright]
typeCheckingMode = "off"
